But how to use older freed blocks? Well, we can mark them as available - which means marking allocated blocks as unavailable.
First of all, we add a new field to our header: a boolean value called
"available". Once a block is created, this field is false, since the block is
in use. Once we call abfree() and the block is not the youngest one, however,
we change this header's field to true.

We also need to know the size of the block. One cannot reuse an 8 size block to
a newly requested 16 size block, for example. So, to know if an available block
is big enough to a newly requested allocation, we also add a field to the
header, named "size", with the size of the block in bytes.

So far, the block will not be reused only by this change. We need to change
abmalloc() yet more, which we should do in the next commit.