Right now, we have a pointer to the topmost, newest block in the "last" variable; from it, we walk back to find a free block. We follow this approach because we incrementally improved our abmalloc() function, but it is not the best one for a bunch of reasons.
For example, consider the following scenario: we allocate four memory blocks,
pointed by "ptr1", "ptr2", "ptr3" and "ptr4" respectively, all of the same
size.

  [####] <- [####] <- [####] <- [####]

We then free "ptr1" and "ptr3".

  [    ] <- [####] <- [    ] <- [####]

Now we allocate a third memory block of the very same size. Since we are
walking backwards, we will get the block once pointed by "ptr3" (instead of
"ptr1"):

  [    ] <- [####] <- [####] <- [####]

If we free "ptr4", we will only reduce the heap by the size of one block:

  [    ] <- [####] <- [####]

Now, if, when searching for free blocks, we walked from the oldest block in the
bottom of the heap towars the newest one, thinks would be different. First, we
would allocate them as before...

  [####] -> [####] -> [####] -> [####]

And then free them:

  [    ] -> [####] -> [    ] -> [####]

However, when looking for a new block to allocate, we would find the
oldest one:

  [####] -> [####] -> [    ] -> [####]

Now, when we free the last block, we will have _two_ free blocks on the top of
the heap, so we can reduce the heap in twice the block size:

  [####] -> [####]

Of course, this is a contrived example but it points to a reall issue: by
re-using the newest available blocks, we are accumulating available blocks at
the bottom of the heap. This is memory that could be used, and will not be, and
the heap will keep growing.

We do not want this to happen, so we wrote the test here commited to ensure the
second behavior.