So, now we have a variable to point out to the first block, and a field in each header pointing to the next. We just need to iterate from the first one to the last, until finding an available block of appropriate size.
To simplify things a bit, we are going to separate two different cases where we
need to create a block. In the first one, there is no block at all yet: in this
case, first is NULL and we just create the block at once. To handle this case
beforehand will make the rest easier.

If there are blocks already, we iterate over them until finding a suitable one.
The difference is that we start from the first one, and we go the the next one,
instead of the previous one. If we find a suitable block, we mark it as
unavailable and return it, as we already do. If no, we extend the heap,
initialize the new block and return it. It is important to not forget to set
the next field from the previous last block!

No much changes with abfree(): we just need to ensure, when releasing the last
block and reducing the heap, that the new last block will have its last field
nullified.